Move long tests in test_ham_unrestricted_chemical
